+++ core/src/main/java/org/elasticsearch/search/suggest/completion/CompletionSuggestion.java
@@ -173,7 +173,7 @@ public final class CompletionSuggestion extends Suggest.Suggestion<CompletionSug
             }
 
             @Override
-            public void setScore(float score) {
+            public void setScore(double score) {
                 super.setScore(score);
             }
